Remove crate level re-export of gps public items
These are nearly implementation details so I don't need to
re-export them. The main benefit of exporting them is
if someone wanted to try and implement an external handler
for something that needs to interact with GPS coords.
I'm not even sure if I export the right stuff for an external
service handler to be possible as the code stands anyways.